+++ CHANGES.md
@@ -4,6 +4,9 @@ Full changelog
 v0.6 (unreleased)
 -----------------
 
+* Prevent axes from moving around when data viewers are being resized, and
+  instead make the absolute margins between axes and figure edge fixed. [#745]
+
 * When a box has been drawn to extract a spectrum from a cube, the box can
   then be moved by pressing the control key and dragging it. [#707]